+++ apps/mrpt_apps_gui/ReactiveNavigationDemo/reactive_navigator_demoMain.cpp
@@ -1337,7 +1337,8 @@ void reactive_navigator_demoframe::simulateOneStep(double time_step)
         const size_t N_STEPS = 20;
         for (size_t j = 0; j < N_STEPS; j++)
         {
-          const double sec = log->gaps_ini[i] + j * (log->gaps_end[i] - log->gaps_ini[i]) /
+          const double sec = log->gaps_ini[i] + static_cast<double>(j) *
+                                                    (log->gaps_end[i] - log->gaps_ini[i]) /
                                                     static_cast<double>(N_STEPS - 1);
           const double ang = M_PI * (-1.0 + 2.0 * sec / static_cast<double>(nObs));
 
@@ -1356,10 +1357,11 @@ void reactive_navigator_demoframe::simulateOneStep(double time_step)
       for (size_t i = 0; i <= nObs; i++)
       {
         const double d0 = lfr.infoPerPTG[sel_PTG].TP_Obstacles[i % nObs];
-        const double a0 = M_PI * (-1.0 + 2.0 * ((i % nObs) + 0.5) / static_cast<double>(nObs));
+        const double a0 =
+            M_PI * (-1.0 + 2.0 * (static_cast<double>(i % nObs) + 0.5) / static_cast<double>(nObs));
         const double d1 = lfr.infoPerPTG[sel_PTG].TP_Obstacles[(i + 1) % nObs];
-        const double a1 =
-            M_PI * (-1.0 + 2.0 * (((i + 1) % nObs) + 0.5) / static_cast<double>(nObs));
+        const double a1 = M_PI * (-1.0 + 2.0 * (static_cast<double>((i + 1) % nObs) + 0.5) /
+                                             static_cast<double>(nObs));
         gl_tp_obstacles->appendLine(
             d0 * cos(a0), d0 * sin(a0), 0.0, d1 * cos(a1), d1 * sin(a1), 0.0);
       }
