+++ apps/mrpt_apps_gui/robot-map-gui/CUndoManager.cpp
@@ -20,13 +20,13 @@ void CUndoManager::addAction(UserAction undo, UserAction redo)
 
   m_undo.push_back(undo);
   m_redo.push_back(redo);
-  m_lastAction = m_undo.size() - 1;
+  m_lastAction = static_cast<int>(m_undo.size()) - 1;
 }
 
 UserAction CUndoManager::redoAction()
 {
   ASSERT_(!m_redo.empty());
-  int redoSize = m_redo.size();
+  int redoSize = static_cast<int>(m_redo.size());
   ASSERT_(m_lastAction < redoSize);
   ++m_lastAction;
   auto it = m_redo.begin() + m_lastAction;
@@ -39,7 +39,7 @@ UserAction CUndoManager::redoAction()
 bool CUndoManager::hasRedo() const
 {
   bool f = !m_redo.empty();
-  int redoLastIndex = m_redo.size() - 1;
+  int redoLastIndex = static_cast<int>(m_redo.size()) - 1;
   bool s = m_lastAction < redoLastIndex;
   return (f && s);
 }
@@ -47,7 +47,7 @@ bool CUndoManager::hasRedo() const
 UserAction CUndoManager::undoAction()
 {
   ASSERT_(!m_undo.empty());
-  int undoSize = m_undo.size();
+  int undoSize = static_cast<int>(m_undo.size());
   ASSERT_(m_lastAction < undoSize);
   auto it = m_undo.begin() + m_lastAction;
   ASSERT_(it != m_undo.end());
