+++ modules/mrpt_hwdrivers/src/CSICKTim561Eth_2050101.cpp
@@ -436,8 +436,8 @@ bool CSICKTim561Eth::decodeScan(char* buff, CObservation2DRangeScan& outObservat
   outObservation.rightToLeft = false;
   outObservation.stdError = 0.012f;
   outObservation.sensorPose = m_sensorPose;
-  outObservation.beamAperture = m_beamApperture;
-  outObservation.maxRange = m_maxRange;
+  outObservation.beamAperture = static_cast<float>(m_beamApperture);
+  outObservation.maxRange = static_cast<float>(m_maxRange);
   outObservation.timestamp = mrpt::Clock::now();
   outObservation.sensorLabel = m_sensorLabel;
